first part of structure from motion.
Recovery of relative pose between two camera images.
Triangulation of 3D points.

+++ source/system-config.hpp
@@ -1,6 +1,6 @@
 #pragma once
 namespace mvSLAM
 {
-using ScalarType = double;
+using ScalarType = float;
 const ScalarType epsilon = (ScalarType) 1e-4;
 }
